Option::make renamed to Option::ensure Written test for Closure returned from Closure, passed to Option::ensure
